[skip ci] TECH-144 change goosies 2023 -> 2024 and goosies 2024 -> 2025 (#129)
* change goosies 2023 -> 2024 and goosies 2024 -> 2025
