+++ public/img/goosies/2025photos/ccdr 2025-04-25 192550.801.jpeg
