+++ src/app/events/goosies-2025/page.tsx
@@ -1,0 +1,12 @@
+import EventPhotoPage from '@/components/event-photo-page';
+
+export default function Goosies2025() {
+    return (
+        <EventPhotoPage
+            eventKey="goosies-2025"
+            photoDir="/img/goosies/2025photos/"
+            headerImageIdx={0}
+            keepHeaderImageInBody={true}
+        />
+    );
+}
